requirements.txt
# DNSScience Utility - Requirements
# Python 3.8+ required

# Core DNS library
dnspython>=2.4.0

# HTTP library for DoH support
requests>=2.31.0

# YAML configuration support
PyYAML>=6.0

# Optional dependencies for advanced features

# For visualization (heatmaps, charts)
matplotlib>=3.7.0
numpy>=1.24.0

# For enhanced output formatting
colorama>=0.4.6

# For testing (development)
pytest>=7.4.0
pytest-cov>=4.1.0

# For code quality (development)
pylint>=2.17.0
black>=23.3.0
mypy>=1.3.0

# For documentation (development)
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0

